Move high-level interaction-related code out of the renderer.
