+++ src/renderer/webgl/WebGLRenderer.ts
@@ -15,17 +15,14 @@ import {
     VertexBuffer,
 } from "picogl";
 
-import { OsrsMenuEntry } from "../../components/rs/menu/OsrsMenu";
 import { createTextureArray } from "./PicoTexture";
-import { MenuTargetType } from "../../rs/MenuEntry";
 import { Scene } from "../../rs/scene/Scene";
-import { isTouchDevice, isWebGL2Supported } from "../../util/DeviceUtil";
+import { isWebGL2Supported } from "../../util/DeviceUtil";
 import { MapViewer } from "../../mapviewer/MapViewer";
 import { MapViewerRenderer } from "../../mapviewer/MapViewerRenderer";
 import { RendererType, WEBGL } from "../Renderers";
 import { DrawRange, NULL_DRAW_RANGE } from "../DrawRange";
-import { InteractType } from "../InteractType";
-import { Interactions } from "../Interactions";
+import { INTERACTION_RADIUS, INTERACT_BUFFER_COUNT, Interactions } from "../Interactions";
 import { WebGLMapSquare } from "./WebGLMapSquare";
 import { SdMapData } from "../loader/SdMapData";
 import { SdMapDataLoader } from "../loader/SdMapDataLoader";
@@ -41,8 +38,6 @@ import { CacheLoaders } from "../../rs/cache/CacheLoaders";
 const MAX_TEXTURES = 2048;
 const TEXTURE_SIZE = 128;
 
-const INTERACT_BUFFER_COUNT = 2;
-const INTERACTION_RADIUS = 5;
 
 interface ColorRgb {
     r: number;
@@ -742,7 +737,7 @@ export class WebGLRenderer extends MapViewerRenderer<WebGLMapSquare> {
         const currInteractions = this.interactions[frameCount % this.interactions.length];
 
         if (!inputManager.isPointerLock()) {
-            this.checkInteractions(currInteractions);
+            this.prepareInteractions(currInteractions);
         } else if (this.hoveredMapIds.size > 0) {
             this.hoveredMapIds.clear();
         }
@@ -1090,7 +1085,7 @@ export class WebGLRenderer extends MapViewerRenderer<WebGLMapSquare> {
         }
     }
 
-    checkInteractions(interactions: Interactions): void {
+    prepareInteractions(interactions: Interactions): void {
         const interactReady = interactions.check(
             this.gl,
             this.hoveredMapIds,
@@ -1104,186 +1099,7 @@ export class WebGLRenderer extends MapViewerRenderer<WebGLMapSquare> {
             return;
         }
 
-        const frameCount = this.stats.frameCount;
-
-        const inputManager = this.mapViewer.inputManager;
-        const isMouseDown = inputManager.dragX !== -1 || inputManager.dragY !== -1;
-        const picked = inputManager.pickX !== -1 && inputManager.pickY !== -1;
-
-        if (!interactReady && !picked) {
-            return;
-        }
-
-        const menuCooldown = isTouchDevice ? 50 : 10;
-
-        if (
-            inputManager.mouseX === -1 ||
-            inputManager.mouseY === -1 ||
-            frameCount - this.mapViewer.menuOpenedFrame < menuCooldown
-        ) {
-            return;
-        }
-
-        // Don't auto close menu on touch devices
-        if (this.mapViewer.menuOpen && !picked && !isMouseDown && isTouchDevice) {
-            return;
-        }
-
-        if (!picked && !this.mapViewer.tooltips) {
-            this.mapViewer.closeMenu();
-            return;
-        }
-
-        const menuEntries: OsrsMenuEntry[] = [];
-        const examineEntries: OsrsMenuEntry[] = [];
-
-        const locIds = new Set<number>();
-        const objIds = new Set<number>();
-        const npcIds = new Set<number>();
-
-        for (let i = 0; i < INTERACTION_RADIUS + 1; i++) {
-            const indices = this.closestInteractIndices.get(i);
-            if (!indices) {
-                continue;
-            }
-            for (const index of indices) {
-                const interactId = this.interactBuffer[index];
-                const interactType = this.interactBuffer[index + 2];
-                if (interactType === InteractType.LOC) {
-                    const locType = this.cacheLoaders.locTypeLoader.load(interactId);
-                    if (locType.name === "null" && !this.mapViewer.debugId) {
-                        continue;
-                    }
-                    if (locIds.has(interactId)) {
-                        continue;
-                    }
-                    locIds.add(interactId);
-
-                    for (const option of locType.actions) {
-                        if (!option) {
-                            continue;
-                        }
-                        menuEntries.push({
-                            option,
-                            targetId: locType.id,
-                            targetType: MenuTargetType.LOC,
-                            targetName: locType.name,
-                            targetLevel: -1,
-                            onClick: this.mapViewer.closeMenu,
-                        });
-                    }
-
-                    examineEntries.push({
-                        option: "Examine",
-                        targetId: locType.id,
-                        targetType: MenuTargetType.LOC,
-                        targetName: locType.name,
-                        targetLevel: -1,
-                        onClick: this.mapViewer.onExamine,
-                    });
-                } else if (interactType === InteractType.OBJ) {
-                    const objType = this.cacheLoaders.objTypeLoader.load(interactId);
-                    if (objType.name === "null" && !this.mapViewer.debugId) {
-                        continue;
-                    }
-                    if (objIds.has(interactId)) {
-                        continue;
-                    }
-                    objIds.add(interactId);
-
-                    for (const option of objType.groundActions) {
-                        if (!option) {
-                            continue;
-                        }
-                        menuEntries.push({
-                            option,
-                            targetId: objType.id,
-                            targetType: MenuTargetType.OBJ,
-                            targetName: objType.name,
-                            targetLevel: -1,
-                            onClick: this.mapViewer.closeMenu,
-                        });
-                    }
-
-                    examineEntries.push({
-                        option: "Examine",
-                        targetId: objType.id,
-                        targetType: MenuTargetType.OBJ,
-                        targetName: objType.name,
-                        targetLevel: -1,
-                        onClick: this.mapViewer.onExamine,
-                    });
-                } else if (interactType === InteractType.NPC) {
-                    let npcType = this.cacheLoaders.npcTypeLoader.load(interactId);
-                    if (npcType.transforms) {
-                        const transformed = npcType.transform(
-                            this.cacheLoaders.varManager,
-                            this.cacheLoaders.npcTypeLoader,
-                        );
-                        if (!transformed) {
-                            continue;
-                        }
-                        npcType = transformed;
-                    }
-                    if (npcType.name === "null" && !this.mapViewer.debugId) {
-                        continue;
-                    }
-                    if (npcIds.has(interactId)) {
-                        continue;
-                    }
-                    npcIds.add(interactId);
-
-                    for (const option of npcType.actions) {
-                        if (!option) {
-                            continue;
-                        }
-                        menuEntries.push({
-                            option,
-                            targetId: npcType.id,
-                            targetType: MenuTargetType.NPC,
-                            targetName: npcType.name,
-                            targetLevel: npcType.combatLevel,
-                            onClick: this.mapViewer.closeMenu,
-                        });
-                    }
-
-                    examineEntries.push({
-                        option: "Examine",
-                        targetId: npcType.id,
-                        targetType: MenuTargetType.NPC,
-                        targetName: npcType.name,
-                        targetLevel: npcType.combatLevel,
-                        onClick: this.mapViewer.onExamine,
-                    });
-                }
-            }
-        }
-
-        menuEntries.push({
-            option: "Walk here",
-            targetId: -1,
-            targetType: MenuTargetType.NONE,
-            targetName: "",
-            targetLevel: -1,
-            onClick: this.mapViewer.closeMenu,
-        });
-        menuEntries.push(...examineEntries);
-        menuEntries.push({
-            option: "Cancel",
-            targetId: -1,
-            targetType: MenuTargetType.NONE,
-            targetName: "",
-            targetLevel: -1,
-            onClick: this.mapViewer.closeMenu,
-        });
-
-        this.mapViewer.menuOpen = picked;
-        if (picked) {
-            this.mapViewer.menuOpenedFrame = frameCount;
-        }
-        this.mapViewer.menuX = inputManager.mouseX;
-        this.mapViewer.menuY = inputManager.mouseY;
-        this.mapViewer.menuEntries = menuEntries;
+        this.checkInteractions(interactReady, this.interactBuffer, this.closestInteractIndices)
     }
 
     override async cleanUp(): Promise<void> {
